Added user subcommand to the CLI for managing users and implemented functions that create users: main.py user create test_user "Test User" "test@user.dev"
and setting their area of interest/research:

main.py user set-area-of-interest test_user "Leveraging LLMs for Personalized Suggestion and Summarization of Scientific Publications"

The latter is meant to take a user's natural language description of what area they would like to get updates for and find a set of OpenAlex topics that best match that description. To achieve that, the description is first aligned structurally with the format used for creating topic embeddings. This improves similarity search. Then, the aligned description is embedded and used to retrieve the 5 most relevant topics from the database using cosine similarity. These topics are then saved to the user configuration.